+++ subsys/llext/CMakeLists.txt
@@ -7,6 +7,8 @@ if(CONFIG_LLEXT)
   zephyr_library_sources(
 		llext.c
 		llext_mem.c
+		llext_load.c
+		llext_link.c
 		llext_export.c
 		buf_loader.c
 	)
