+++ subsys/llext/llext_priv.h
@@ -39,4 +39,25 @@ static inline void llext_free(void *ptr)
 	k_heap_free(&llext_heap, ptr);
 }
 
+/*
+ * ELF parsing (llext_load.c)
+ */
+
+int do_llext_load(struct llext_loader *ldr, struct llext *ext,
+		  struct llext_load_param *ldr_parm);
+
+elf_shdr_t *llext_section_by_name(struct llext_loader *ldr, const char *search_name);
+
+static inline const char *llext_string(struct llext_loader *ldr, struct llext *ext,
+				       enum llext_mem mem_idx, unsigned int idx)
+{
+	return (char *)ext->mem[mem_idx] + idx;
+}
+
+/*
+ * Relocation (llext_link.c)
+ */
+
+int llext_link(struct llext_loader *ldr, struct llext *ext, bool do_local);
+
 #endif /* ZEPHYR_SUBSYS_LLEXT_PRIV_H_ */
